Se modifica estructura de bd y se agregan queries para agregar datos default y se define permiso como enum estatico

+++ tp-cuatrimetral-equipo-2A/dominio/Venta.cs
@@ -9,7 +9,7 @@ namespace dominio
     {
         public int ID { get; set; }
         public Usuario Usuario { get; set; }
-        public int SumaTotal { get; set; }
+        public float SumaTotal { get; set; }
         public DateTime FechaVenta { get; set; }
     }
 }
